Refactor volume code to separate class
